# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API simulador de escuela",
        "description": "Api donde podras ver, editar, crear y eliminar los datos de los estudiantes, profesores y materias, asi como tambien matricular un alumno a una materia.",
        "contact": {
            "email": "arturo@arturo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/asistencias": {
            "get": {
                "tags": [
                    "Asistencias"
                ],
                "summary": "Listado de todos los registros de los alumnos",
                "operationId": "c2e86616d8002ab97af5e8a753b8ed37",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "alumno_id": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "fecha": {
                                                        "type": "string",
                                                        "example": "2024/01/25"
                                                    },
                                                    "asistencia": {
                                                        "type": "string",
                                                        "example": "A"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Asistencias"
                ],
                "summary": "Registrar la asistencia de un alumno",
                "operationId": "007ddf3bd33265e9691e3724e4fecef5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "alumno_id": {
                                                "type": "number"
                                            },
                                            "fecha": {
                                                "type": "string"
                                            },
                                            "asistencia": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "alumno_id": "1",
                                    "fecha": "2024/02/24",
                                    "asistencia": "A"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "alumno_id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "fecha": {
                                            "type": "string",
                                            "example": "2024/02/24"
                                        },
                                        "asistencia": {
                                            "type": "string",
                                            "example": "A"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The asistencias field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}